Initial Commit of Email Reminder Tool
Last we were discussing how to simplify our check-in meetings, and
we had the idea of just having a script prompt use to update issue
offline, rather than doing it at our checkin meeting.

This is the initial attempt at that. It take a config file that
defines intervals at which we want people to be prodded, and then
sends them, and potentially a pm or manager, reminders if an issue has
been sitting for a while. It user interface is similar to the report
tool, where is takes the github token, target repo and organization
as arguements.

There are chunks of this that should be refactored into smaller
functions, and I want to add usage of bunches. However, I have
already done too much for one commit.