Refactor for code reuse and adding of assignees
This commit refactored the decorated issue generation process into
two steps: filtration of marked issues and conversion to bunches.
This was important as the bunching/decoration process is needed
by report.py and the email notification functionality, but the
fitration logic is different.

I also changed the assignee column in the shared table preperation
logic to show assignees instead of the assignee. When there was
multiple assignees, the listed assignee for an issue was sometimes
the person who had already passed it on. Granted, the best practice
is to update these as soon as an issues responsibility changes
but for the purposes of automated report I thought it was better
to list all the people that are currently assigned to something.