dynamic config loc and bunches
I added a 4th optional arguement to be able to specify the config.yaml
fullpath at run time. This was pretty simple. At some point, we
should add schema validation of the yaml file though, so the
code does not at erratically in response to mis-spellings or
formatting in that file.

I also change the timeframe structure into a bunch to slightly simplify the
syntax at the lower levels.