Upgraded arg parser and added default debug mode
As this system could shot out email to a lot of semi-random people,
I decided to change its default to a sort of safe mode. If you run
it with just the required arguements, it will print out the emails
it is thinking of sending. If you add --send, it will send out the
emails. To do this, I upgrade the arg parse from clint to argpars
which has also enabled some type validation.