Adding triage label and catch-up to catchup
When there is no recipent for an issue, the label
triage will now be added to it. This works with
our current process. In the long run, this script
may want to support an email address entry that
could get a heads up for these.

Also, I changed catch-up to catchup, as I'm not
sure the former worked with bunches.