Adding days_multiplier import and enhancing debugs
I added the days multiplier import to the yaml file. This is to aide
catching-up and getting the email system up to date. In our case, our
final standards are much higher than where we are now, but we don't
want people's first interaction with this tool to be getting 100
notifications. Therefore, I'm adding this days_multiplier import
which I can start out at 10x, and then slowly wind down to 1x / gone
once were are at our target level.