adding email override list
There is a flaw in my plan. The email that the google API provides
you is someones public email. This has no connection
to the email that a person gets their notifications sent
to, and in fact that other email is not accessible.
This makes sense a person with notoriety would not
want the public having their private email.

There are two solutions: have everyone change their
publicly facing email address to rentjungle.com or
add a software fix. I got resistance from teh team,
so i just went with the later.

This addition is a quick fix, to get this in production
that additional piping needed to get this email override
down to the recipient list creator has convinced me
that this whole this is coupled enough that it should
just be a large object.