changing and fixing formating of table
the same table logic is used in both out maintenance
emails as well as the automatic reminder emails, and both s
seem a bit wide for some screens. Also, I think some of the
column widths coudl be better. First, I reduced the overall
table width and the padding around text in it. Second, I
replaced the width entries in th with style:width entries
in the th entries, as the width values are ignored in html5
.